+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Airbrake Changelog
 
 * Rails: added support for Curb for HTTP performance breakdown
   ([#957](https://github.com/airbrake/airbrake/pull/957))
+* Rails: added support for Excon for HTTP performance breakdown
+  ([#958](https://github.com/airbrake/airbrake/pull/958))
 
 ### [v9.1.0][v9.1.0] (April 17, 2019)